+++ lib/promrelabel/debug_test.go
@@ -10,10 +10,10 @@ import (
 
 // TestWriteRelabelDebugSupportFormats verifies the relabeling debug input, rules and output.
 func TestWriteRelabelDebugSupportFormats(t *testing.T) {
-	f := func(input, rule, expect string) {
+	f := func(input, rules, expect string) {
 		// execute
 		outputWriter := bytes.NewBuffer(nil)
-		writeRelabelDebug(outputWriter, false, "", input, rule, "json", nil)
+		writeRelabelDebug(outputWriter, false, "", input, rules, 0, 0, "json", nil)
 
 		// the response is in JSON with HTML content, extract the `resultingLabels` in JSON and unescape it.
 		resultingLabels := fastjson.GetString(outputWriter.Bytes(), `resultingLabels`)
@@ -41,4 +41,20 @@ func TestWriteRelabelDebugSupportFormats(t *testing.T) {
 	f(`{_name__="metric_name"`, ruleTestParsing, ``)
 	f(`_name__="metric_name}"`, ruleTestParsing, ``)
 	f(`metrics_name}"`, ruleTestParsing, ``)
+
+	// test multiple rules including remote writes
+	// drop all labels and add one in URL relabeling
+	rule1 := `
+- action: labeldrop
+  regex: "drop_me_metrics_relabel"
+`
+	rule2 := `
+- action: labeldrop
+  regex: "drop_me_remote_write_relabel"
+`
+	rule3 := `
+- target_label: add_me_url_relabel
+  replacement: added
+`
+	f(`{__name__="metric_name", drop_me_metrics_relabel="1", drop_me_remote_write_relabel="2"}`, rule1+rule2+rule3, `metric_name{add_me_url_relabel="added"}`)
 }